Most of frirst review
