+++ examples/Calc/CMakeLists.txt
@@ -15,7 +15,7 @@ if(Boost_UNIT_TEST_FRAMEWORK_FOUND)
     target_link_libraries(BoostCalculatorSteps Calc ${CUKE_LIBRARIES} ${Boost_UNIT_TEST_FRAMEWORK_LIBRARY})
 endif()
 
-if(QT_LIBRARIES)
+if(QT5_FOUND)
     add_executable(QtTestCalculatorSteps features/step_definitions/QtTestCalculatorSteps)
     target_link_libraries(QtTestCalculatorSteps Calc ${QT_LIBRARIES} ${CUKE_LIBRARIES})
 endif()
