+++ examples/CalcQt/features/step_definitions/QtTestCalculatorQtSteps.cpp
@@ -1,8 +1,11 @@
 #include <QTest>
-#include <cucumber-cpp/autodetect.hpp>
+#include "cucumber-cpp/autodetect.hpp"
 #include <cstdlib>
 #include <QApplication>
 #include "CalculatorWidget.h"
+
+std::istream& operator>> (std::istream& in, QString& val) { std::string s; in >> s; val = QString::fromLocal8Bit(s.c_str()); return in; }
+std::ostream& operator<< (std::ostream& out, const QString& val) { out << val.toLocal8Bit().data(); return out; }
 
 static int argc = 0;
 static QApplication app(argc, 0);
@@ -12,13 +15,14 @@ int millisecondsToWait() {
     if (milliseconds < 0)
     {
         char* envVariable = getenv("CALCQT_STEP_DELAY");
-        milliseconds = (0 != envVariable) ? atoi(envVariable) : 0;
+        milliseconds = (envVariable) ? atoi(envVariable) : 0;
     }
     return milliseconds;
 }
 
-std::istream& operator>> (std::istream& in, QString& val) { std::string s; in >> s; val = s.c_str(); return in; }
-std::ostream& operator<< (std::ostream& out, const QString& val) { out << val.toLatin1().data(); return out; }
+AFTER_STEP(){
+    QTest::qWait(millisecondsToWait());
+}
 
 GIVEN("^I just turned on the calculator$") {
     cucumber::ScenarioScope<CalculatorWidget> calculator;
@@ -29,44 +33,41 @@ GIVEN("^I just turned on the calculator$") {
 #else
     QTest::qWaitForWindowShown(calculator.get());
 #endif
-    QTest::qWait(millisecondsToWait());
 }
 
 WHEN("^I press (\\d+)$") {
     REGEX_PARAM(unsigned int, n);
     cucumber::ScenarioScope<CalculatorWidget> calculator;
-    QTest::keyClick(calculator.get(), Qt::Key_0 + n, Qt::NoModifier, millisecondsToWait());
+    QTest::keyClick(calculator.get(), Qt::Key_0 + n, Qt::NoModifier, 0);
 }
 
 WHEN("^I press add") {
     cucumber::ScenarioScope<CalculatorWidget> calculator;
-    QTest::keyClick(calculator.get(), Qt::Key_Plus, Qt::NoModifier, millisecondsToWait());
+    QTest::keyClick(calculator.get(), Qt::Key_Plus, Qt::NoModifier, 0);
 }
 
 WHEN("^I press calculate") {
     cucumber::ScenarioScope<CalculatorWidget> calculator;
-    QTest::keyClick(calculator.get(), Qt::Key_Return, Qt::NoModifier, millisecondsToWait());
+    QTest::keyClick(calculator.get(), Qt::Key_Return, Qt::NoModifier, 0);
 }
 
 WHEN("^I press clear") {
     cucumber::ScenarioScope<CalculatorWidget> calculator;
-    QTest::keyClick(calculator.get(), Qt::Key_Escape, Qt::NoModifier, millisecondsToWait());
+    QTest::keyClick(calculator.get(), Qt::Key_Escape, Qt::NoModifier, 0);
 }
 
 WHEN("^I press subtract") {
     cucumber::ScenarioScope<CalculatorWidget> calculator;
-    QTest::keyClick(calculator.get(), Qt::Key_Minus, Qt::NoModifier, millisecondsToWait());
+    QTest::keyClick(calculator.get(), Qt::Key_Minus, Qt::NoModifier, 0);
 }
 
 THEN("^the display should be empty$") {
     cucumber::ScenarioScope<CalculatorWidget> calculator;
     QCOMPARE(0, calculator->display().size());
-    QTest::qWait(millisecondsToWait());
 }
 
 THEN("^the display should show (.*)$") {
     REGEX_PARAM(QString, expected);
     cucumber::ScenarioScope<CalculatorWidget> calculator;
     QCOMPARE(calculator->display(), expected);
-    QTest::qWait(millisecondsToWait());
 }
