+++ src/CMakeLists.txt
@@ -19,6 +19,7 @@ if(QT5_FOUND)
     qt5_wrap_cpp(MOC_FILE ../include/cucumber-cpp/internal/drivers/QtTestDriver.hpp)
     list(APPEND CUKE_SOURCES drivers/QtTestDriver.cpp)
     list(APPEND CUKE_SOURCES ${MOC_FILE})
+    list(APPEND CUKE_DEP_LIBRARIES ${QT_LIBRARIES})
 endif()
 
 if(GTEST_FOUND)
